requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
filelock==3.12.2
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
platformdirs==3.10.0
pycodestyle==2.11.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
tk==0.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.2
Werkzeug==2.3.7
yarl==1.9.2